spec: move powerSaveBlocker spec to main process (#19236)
